pyproject.toml
[project]
name = "autodoc-adk"
version = "0.1.0"
description = "AI-powered documentation generator using Google ADK, Prefect, and FastAPI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "google-adk>=0.3.0",
    "prefect>=3.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "pydantic-settings>=2.0",
    "opentelemetry-api>=1.28.0",
    "opentelemetry-sdk>=1.28.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.28.0",
    "opentelemetry-instrumentation-logging>=0.49b0",
    "litellm>=1.50.0",
    "fastmcp>=2.0.0",
    "httpx>=0.28.0",
    "boto3>=1.35.0",
    "alembic>=1.14.0",
    "pyyaml>=6.0",
    "python-json-logger>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ── Ruff ──

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "SIM", # flake8-simplify
    "T20", # flake8-print
    "RUF", # ruff-specific rules
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults (needed for FastAPI Depends)
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# ── Pytest ──

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = [
    "ignore::DeprecationWarning",
]
markers = [
    "integration: integration tests exercising Prefect flow orchestration",
]
